# src/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "2rist API Doc",
    "description": "API documentation for 2rist app.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "tags": [
    {
      "name": "Hotels",
      "description": "Endpoints related to hotels"
    },
    {
      "name": "Auth",
      "description": "Endpoints related to authentication"
    },
    {
      "name": "Feedbacks",
      "description": "Endpoints related to user feedback"
    }
  ],
  "paths": {
    "/auth/google": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/auth/google/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/auth/apple": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/auth/apple/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/protected": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/hotels": {
      "post": {
        "tags": [
          "Hotels"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "search": {
                    "example": "any"
                  },
                  "checkInYear": {
                    "example": "any"
                  },
                  "checkInMonth": {
                    "example": "any"
                  },
                  "checkInDay": {
                    "example": "any"
                  },
                  "checkOutYear": {
                    "example": "any"
                  },
                  "checkOutMonth": {
                    "example": "any"
                  },
                  "checkOutDay": {
                    "example": "any"
                  },
                  "adultCount": {
                    "example": "any"
                  },
                  "childCount": {
                    "example": "any"
                  },
                  "roomCount": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/hotel/{id}": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "hotel id"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/favorites": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Hotels"
        ],
        "description": "Clear the favorite hotel list of user.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/favorites/{hotelId}": {
      "patch": {
        "tags": [
          "Hotels"
        ],
        "description": "Add or Delete a hotel from the favorite hotel list.",
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Id of the hotel to add to the favorite list."
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/feedbacks": {
      "post": {
        "tags": [
          "Feedbacks"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "message": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}